+++ tests/puppeteer/campaigns/data/faker/webservice.js
@@ -2,6 +2,7 @@ const faker = require('faker');
 
 module.exports = class Webservice {
   constructor(webserviceToCreate = {}) {
+    this.key = webserviceToCreate.key || faker.random.uuid().substring(0,32);
     this.keyDescription = webserviceToCreate.keyDescription || faker.lorem.sentence();
     this.status = webserviceToCreate.status === undefined ? true : webserviceToCreate.status;
   }
