+++ tests/puppeteer/pages/BO/advancedParameters/webservice/add.js
@@ -9,6 +9,7 @@ module.exports = class AddWebserviceKey extends BOBasePage {
     this.pageTitleEdit = 'Webservice •';
 
     // Selectors
+    this.webserviceKeyInput = '#webservice_key_key';
     this.generateButton = 'button.js-generator-btn';
     this.keyDescriptionTextarea = '#webservice_key_description';
     this.statusSwitchLabel = 'label[for="webservice_key_status_%ID"]';
@@ -22,10 +23,12 @@ module.exports = class AddWebserviceKey extends BOBasePage {
   /**
    * Fill form for add/edit webservice key
    * @param webserviceData
+   * @param toGenerate
    * @return {Promise<textContent>}
    */
-  async createEditWebservice(webserviceData) {
-    await this.page.click(this.generateButton);
+  async createEditWebservice(webserviceData, toGenerate = true) {
+    if (toGenerate) await this.page.click(this.generateButton);
+    else await this.setValue(this.webserviceKeyInput, webserviceData.key);
     await this.setValue(this.keyDescriptionTextarea, webserviceData.keyDescription);
     // replace %ID by 1 in the selector if active = YES / 0 if active = NO
     await this.page.click(this.statusSwitchLabel.replace('%ID', webserviceData.status ? 1 : 0));
